Refactor to prevent errors in console
