+++ src/components/Header.js
@@ -4,8 +4,8 @@ export default function Header() {
     return (
         <header className="App-header">
             <p>
-            Welcome to the Mendota Water Temp Viewer!
+                Welcome to the Mendota Water Temp Viewer!
             </p>
-      </header>
+        </header>
     );
 };
